--- index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <style>
-        canvas {
-            background: radial-gradient(circle at bottom, navy 0, black 100%);
-        }
-        
-        body {
-            margin: 0;
-            padding: 0;
-            overflow: hidden;
-        }
-    </style>
-</head>
-
-<body>
-    <canvas id="canvas_sonic" width="" height=""></canvas>
-    <script src="./dist/bundle.js"></script>
-</body>
-
-</html>
