--- main.js
@@ -1,7 +1,0 @@
-"use strict";
-
-var _Animation = _interopRequireDefault(require("Animation.js"));
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-new Anaimation().teste();
