+++ src/app/app-routing.module.ts
@@ -1,8 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { GameComponent } from './game/game.component';
 
 
-const routes: Routes = [];
+const routes: Routes = [{
+  path: '**', component: GameComponent
+}];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
