+++ src/app/game/entity/explosion.ts
