+++ src/app/game/entity/interaction.ts
