+++ src/app/game/game.component.html
@@ -1,0 +1,2 @@
+<nav #barStatus></nav>
+<canvas #canvas width="" height=""></canvas>
