+++ src/app/game/game.component.scss
@@ -1,0 +1,8 @@
+canvas {
+    background: radial-gradient(circle at bottom, navy 0, black 100%);
+}
+
+nav {
+    height: 40px;
+    background: #000;
+}
