+++ src/app/game/structure/diraction.ts
