+++ src/app/game/structure/ianimation.ts
