+++ src/app/game/structure/imagens.ts
@@ -1,7 +1,7 @@
 
 export default class Imagens {
     imgs: Array<any> = [];
-    loads: 0;
+    loads = 0;
     addImage(iAnimation: any, path: string) {
       const image = new Image();
       image.src = path;
@@ -13,7 +13,7 @@ export default class Imagens {
         this.imgs.forEach(img => {
             img.onload = () => {
                this.loads++; 
-               if(this.loads == this.imgs.length) {
+               if(this.loads == this.imgs.length - 1) {
                    callback();
                }
             }
